edit tests, test structure, config files, update readme
pre-commit tests pass

remove spin SDK install in action

update poetry lock

last ditch attempt at optional dependency for action

poetry bugs and other nonsense workaround

test change, poetry unable to ignore spin dependencies, can't test spin

importerror catch

import error catch

workaround S3 access in git action test

update readme for setup guide

update pre-commit config
