+++ .github/workflows/on-push.yml
@@ -19,17 +19,6 @@ jobs:
         with:
           python-version: 3.10.12
 
-      - name: spinnaker SDK download & install
-        uses: actions/cache@v2
-        with:
-          path: ~/spinnaker
-          key: spinnaker-3.1.0.79-amd64
-
-      - run: wget https://github.com/zaitra/vesna_experiment/releases/download/0.1/spinnaker-3.1.0.79-amd64-pkg.tar.gz
-      - run: tar -xzvf spinnaker-3.1.0.79-amd64-pkg.tar.gz
-      - run: sudo apt install qt5-default libgomp1
-      - run: sudo sh spinnaker-3.1.0.79-amd64/install_spinnaker.sh
-
       # Cache the installation of Poetry itself, e.g. the next step. This prevents the workflow
       # from installing Poetry every time, which can be slow. Note the use of the Poetry version
       # number in the cache key, and the "-0" suffix: this allows you to invalidate the cache
@@ -39,7 +28,7 @@ jobs:
         uses: actions/cache@v2
         with:
           path: ~/.local
-          key: poetry-1.2.0-0
+          key: poetry-1.5.0-0
 
       # Install Poetry. You could do this manually, or there are several actions that do this.
       # `snok/install-poetry` seems to be minimal yet complete, and really just calls out to
@@ -50,7 +39,7 @@ jobs:
       # cache it.
       - uses: snok/install-poetry@v1
         with:
-          version: 1.2.0
+          version: 1.5.0
           virtualenvs-create: true
           virtualenvs-in-project: true
 
@@ -65,8 +54,8 @@ jobs:
       # Install dependencies. `--no-root` means "install all dependencies but not the project
       # itself", which is what you want to avoid caching _your_ code. The `if` statement
       # ensures this only runs on a cache miss.
-      - run: poetry install --no-interaction --no-root
-        if: steps.cache-deps.outputs.cache-hit != 'true'
+      #- run: poetry install --no-interaction --no-root
+      #  if: steps.cache-deps.outputs.cache-hit != 'true'
 
       # Now install _your_ project. It's a good idea since it fully-exercises the
       # pyproject.toml and makes that if you add things like console-scripts at some point that
@@ -74,4 +63,4 @@ jobs:
       - run: poetry install --no-interaction
 
       # And finally run tests.
-      - run: poetry run pytest
+      - run: poetry run pytest --ignore=tests/test_camera.py --ignore=tests/test_runs.py --ignore=tests/test_file_access.py
