+++ tests/test_cli_args.py
@@ -1,5 +1,6 @@
 import typer
 from typer.testing import CliRunner
+import structlog
 
 import camera_control.__main__ as cc
